fixit: Changes to DocAI Samples to simplify maintenence
- Consolidated process/batch process samples with processor versions
- Update Quickstart to include Processor Creation
- Consolidate Handle Response Samples into a single file to reduce code duplication
  - ~Not sure if I did the region tags completely correct, is there a way to test?~
  - Update: Tested using Devsite staging and it looks correct
- Moved Operations Tests into a single file and removed exception handling in sample
- Add type updates from fixit #9991
