+++ documentai/snippets/handle_response_sample.py
@@ -1,0 +1,394 @@
+# Copyright 2020 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+# [START documentai_process_ocr_document]
+# [START documentai_process_form_document]
+# [START documentai_process_quality_document]
+# [START documentai_process_specialized_document]
+# [START documentai_process_splitter_document]
+from typing import Sequence
+
+from google.api_core.client_options import ClientOptions
+from google.cloud import documentai  # type: ignore
+
+# TODO(developer): Uncomment these variables before running the sample.
+# project_id = "YOUR_PROJECT_ID"
+# location = "YOUR_PROCESSOR_LOCATION" # Format is "us" or "eu"
+# processor_id = "YOUR_PROCESSOR_ID" # Create processor before running sample
+# processor_version = "rc" # Refer to https://cloud.google.com/document-ai/docs/manage-processor-versions for more information
+# file_path = "/path/to/local/pdf"
+# mime_type = "application/pdf" # Refer to https://cloud.google.com/document-ai/docs/file-types for supported file types
+
+
+# [END documentai_process_ocr_document]
+# [END documentai_process_form_document]
+# [END documentai_process_quality_document]
+# [END documentai_process_specialized_document]
+# [END documentai_process_splitter_document]
+
+
+# [START documentai_process_ocr_document]
+def process_document_ocr_sample(
+    project_id: str,
+    location: str,
+    processor_id: str,
+    processor_version: str,
+    file_path: str,
+    mime_type: str,
+) -> None:
+    # Online processing request to Document AI
+    document = process_document(
+        project_id, location, processor_id, processor_version, file_path, mime_type
+    )
+
+    text = document.text
+    print(f"Full document text: {text}\n")
+    print(f"There are {len(document.pages)} page(s) in this document.\n")
+
+    for page in document.pages:
+        print(f"Page {page.page_number}:")
+        print_page_dimensions(page.dimension)
+        print_detected_langauges(page.detected_languages)
+        print_paragraphs(page.paragraphs, text)
+        print_blocks(page.blocks, text)
+        print_lines(page.lines, text)
+        print_tokens(page.tokens, text)
+
+        # Currently supported in version `pretrained-ocr-v1.1-2022-09-12`
+        if page.image_quality_scores:
+            print_image_quality_scores(page.image_quality_scores)
+
+
+def print_page_dimensions(dimension: documentai.Document.Page.Dimension) -> None:
+    print(f"    Width: {str(dimension.width)}")
+    print(f"    Height: {str(dimension.height)}")
+
+
+def print_detected_langauges(
+    detected_languages: Sequence[documentai.Document.Page.DetectedLanguage],
+) -> None:
+    print("    Detected languages:")
+    for lang in detected_languages:
+        code = lang.language_code
+        print(f"        {code} ({lang.confidence:.1%} confidence)")
+
+
+def print_paragraphs(
+    paragraphs: Sequence[documentai.Document.Page.Paragraph], text: str
+) -> None:
+    print(f"    {len(paragraphs)} paragraphs detected:")
+    first_paragraph_text = layout_to_text(paragraphs[0].layout, text)
+    print(f"        First paragraph text: {repr(first_paragraph_text)}")
+    last_paragraph_text = layout_to_text(paragraphs[-1].layout, text)
+    print(f"        Last paragraph text: {repr(last_paragraph_text)}")
+
+
+def print_blocks(blocks: Sequence[documentai.Document.Page.Block], text: str) -> None:
+    print(f"    {len(blocks)} blocks detected:")
+    first_block_text = layout_to_text(blocks[0].layout, text)
+    print(f"        First text block: {repr(first_block_text)}")
+    last_block_text = layout_to_text(blocks[-1].layout, text)
+    print(f"        Last text block: {repr(last_block_text)}")
+
+
+def print_lines(lines: Sequence[documentai.Document.Page.Line], text: str) -> None:
+    print(f"    {len(lines)} lines detected:")
+    first_line_text = layout_to_text(lines[0].layout, text)
+    print(f"        First line text: {repr(first_line_text)}")
+    last_line_text = layout_to_text(lines[-1].layout, text)
+    print(f"        Last line text: {repr(last_line_text)}")
+
+
+def print_tokens(tokens: Sequence[documentai.Document.Page.Token], text: str) -> None:
+    print(f"    {len(tokens)} tokens detected:")
+    first_token_text = layout_to_text(tokens[0].layout, text)
+    first_token_break_type = tokens[0].detected_break.type_.name
+    print(f"        First token text: {repr(first_token_text)}")
+    print(f"        First token break type: {repr(first_token_break_type)}")
+    last_token_text = layout_to_text(tokens[-1].layout, text)
+    last_token_break_type = tokens[-1].detected_break.type_.name
+    print(f"        Last token text: {repr(last_token_text)}")
+    print(f"        Last token break type: {repr(last_token_break_type)}")
+
+
+def print_image_quality_scores(
+    image_quality_scores: documentai.Document.Page.ImageQualityScores,
+) -> None:
+    print(f"    Quality score: {image_quality_scores.quality_score:.1%}")
+    print("    Detected defects:")
+
+    for detected_defect in image_quality_scores.detected_defects:
+        print(f"        {detected_defect.type_}: {detected_defect.confidence:.1%}")
+
+
+# [END documentai_process_ocr_document]
+# [START documentai_process_form_document]
+def process_document_form_sample(
+    project_id: str,
+    location: str,
+    processor_id: str,
+    processor_version: str,
+    file_path: str,
+    mime_type: str,
+) -> None:
+    # Online processing request to Document AI
+    document = process_document(
+        project_id, location, processor_id, processor_version, file_path, mime_type
+    )
+
+    # Read the table and form fields output from the processor
+    # The form processor also contains OCR data. For more information
+    # on how to parse OCR data please see the OCR sample.
+
+    text = document.text
+    print(f"Full document text: {repr(text)}\n")
+    print(f"There are {len(document.pages)} page(s) in this document.")
+
+    # Read the form fields and tables output from the processor
+    for page in document.pages:
+        print(f"\n\n**** Page {page.page_number} ****")
+
+        print(f"\nFound {len(page.tables)} table(s):")
+        for table in page.tables:
+            num_collumns = len(table.header_rows[0].cells)
+            num_rows = len(table.body_rows)
+            print(f"Table with {num_collumns} columns and {num_rows} rows:")
+
+            # Print header rows
+            print("Columns:")
+            print_table_rows(table.header_rows, text)
+            # Print body rows
+            print("Table body data:")
+            print_table_rows(table.body_rows, text)
+
+        print(f"\nFound {len(page.form_fields)} form field(s):")
+        for field in page.form_fields:
+            name = layout_to_text(field.field_name, text)
+            value = layout_to_text(field.field_value, text)
+            print(f"    * {repr(name.strip())}: {repr(value.strip())}")
+
+
+def print_table_rows(
+    table_rows: Sequence[documentai.Document.Page.Table.TableRow], text: str
+) -> None:
+    for table_row in table_rows:
+        row_text = ""
+        for cell in table_row.cells:
+            cell_text = layout_to_text(cell.layout, text)
+            row_text += f"{repr(cell_text.strip())} | "
+        print(row_text)
+
+
+# [END documentai_process_form_document]
+# [START documentai_process_quality_document]
+def process_document_quality_sample(
+    project_id: str,
+    location: str,
+    processor_id: str,
+    processor_version: str,
+    file_path: str,
+    mime_type: str,
+) -> None:
+    # Online processing request to Document AI
+    document = process_document(
+        project_id, location, processor_id, processor_version, file_path, mime_type
+    )
+
+    # Read the quality-specific information from the output from the
+    # Intelligent Document Quality Processor:
+    # https://cloud.google.com/document-ai/docs/processors-list#processor_doc-quality-processor
+    # OCR and other data is also present in the quality processor's response.
+    # Please see the OCR and other samples for how to parse other data in the
+    # response.
+    for entity in document.entities:
+        conf_percent = f"{entity.confidence:.1%}"
+        page_num = str(int(entity.page_anchor.page_refs[0].page) + 1)
+        print(f"\nPage {page_num} has a quality score of {conf_percent}")
+
+        for prop in entity.properties:
+            conf_percent = f"{prop.confidence:.1%}"
+            print(f"    * {prop.type_} score of {conf_percent}")
+
+
+# [END documentai_process_quality_document]
+
+
+# [START documentai_process_specialized_document]
+def process_document_specialized_sample(
+    project_id: str,
+    location: str,
+    processor_id: str,
+    processor_version: str,
+    file_path: str,
+    mime_type: str,
+) -> None:
+    # Online processing request to Document AI
+    document = process_document(
+        project_id, location, processor_id, processor_version, file_path, mime_type
+    )
+
+    # Extract entities from a specialized document
+    # Most specalized processors follow a similar pattern.
+    # For a complete list of processors see:
+    # https://cloud.google.com/document-ai/docs/processors-list
+    #
+    # OCR and other data is also present in the processor's response.
+    # Refer to the OCR samples for how to parse other data in the response.
+
+    print(f"Found {len(document.entities)} entities:")
+    for entity in document.entities:
+        print_entity(entity)
+        # Print Nested Entities (if any)
+        for prop in entity.properties:
+            print_entity(prop)
+
+
+def print_entity(entity: documentai.Document.Entity) -> None:
+    # Fields detected. For a full list of fields for each processor see
+    # the processor documentation:
+    # https://cloud.google.com/document-ai/docs/processors-list
+    key = entity.type_
+
+    # Some other value formats in addition to text are availible
+    # e.g. dates: `entity.normalized_value.date_value.year`
+    text_value = entity.text_anchor.content
+    confidence = entity.confidence
+    normalized_value = entity.normalized_value.text
+    print(f"    * {repr(key)}: {repr(text_value)}({confidence:.1%} confident)")
+
+    if normalized_value:
+        print(f"    * Normalized Value: {repr(normalized_value)}")
+
+
+# [END documentai_process_specialized_document]
+
+
+# [START documentai_process_splitter_document]
+def process_document_splitter_sample(
+    project_id: str,
+    location: str,
+    processor_id: str,
+    processor_version: str,
+    file_path: str,
+    mime_type: str,
+) -> None:
+    # Online processing request to Document AI
+    document = process_document(
+        project_id, location, processor_id, processor_version, file_path, mime_type
+    )
+
+    # Read the splitter output from a document splitter/classifier processor:
+    # e.g. https://cloud.google.com/document-ai/docs/processors-list#processor_procurement-document-splitter
+    # This processor only provides text for the document and information on how
+    # to split the document on logical boundaries. To identify and extract text,
+    # form elements, and entities please see other processors like the OCR, form,
+    # and specalized processors.
+
+    print(f"Found {len(document.entities)} subdocuments:")
+    for entity in document.entities:
+        conf_percent = f"{entity.confidence:.1%}"
+        pages_range = page_refs_to_string(entity.page_anchor.page_refs)
+
+        # Print subdocument type information, if available
+        if entity.type_:
+            print(
+                f"{conf_percent} confident that {pages_range} a '{entity.type_}' subdocument."
+            )
+        else:
+            print(f"{conf_percent} confident that {pages_range} a subdocument.")
+
+
+def page_refs_to_string(
+    page_refs: Sequence[documentai.Document.PageAnchor.PageRef],
+) -> str:
+    """Converts a page ref to a string describing the page or page range."""
+    if len(page_refs) == 1:
+        num = str(int(page_refs[0].page) + 1)
+        return f"page {num} is"
+
+    nums = ""
+    for page_ref in page_refs:
+        nums += f"{int(page_ref.page) + 1}, "
+    return f"pages {nums[:-2]} are"
+
+
+# [END documentai_process_splitter_document]
+
+
+# [START documentai_process_ocr_document]
+# [START documentai_process_form_document]
+# [START documentai_process_quality_document]
+# [START documentai_process_specialized_document]
+# [START documentai_process_splitter_document]
+def process_document(
+    project_id: str,
+    location: str,
+    processor_id: str,
+    processor_version: str,
+    file_path: str,
+    mime_type: str,
+) -> documentai.Document:
+    # You must set the `api_endpoint` if you use a location other than "us".
+    client = documentai.DocumentProcessorServiceClient(
+        client_options=ClientOptions(
+            api_endpoint=f"{location}-documentai.googleapis.com"
+        )
+    )
+
+    # The full resource name of the processor version, e.g.:
+    # `projects/{project_id}/locations/{location}/processors/{processor_id}/processorVersions/{processor_version_id}`
+    # You must create a processor before running this sample.
+    name = client.processor_version_path(
+        project_id, location, processor_id, processor_version
+    )
+
+    # Read the file into memory
+    with open(file_path, "rb") as image:
+        image_content = image.read()
+
+    # Configure the process request
+    request = documentai.ProcessRequest(
+        name=name,
+        raw_document=documentai.RawDocument(content=image_content, mime_type=mime_type),
+    )
+
+    result = client.process_document(request=request)
+
+    # For a full list of `Document` object attributes, reference this page:
+    # https://cloud.google.com/document-ai/docs/reference/rest/v1/Document
+    return result.document
+
+
+# [END documentai_process_quality_document]
+# [END documentai_process_specialized_document]
+# [END documentai_process_splitter_document]
+def layout_to_text(layout: documentai.Document.Page.Layout, text: str) -> str:
+    """
+    Document AI identifies text in different parts of the document by their
+    offsets in the entirety of the document"s text. This function converts
+    offsets to a string.
+    """
+    response = ""
+    # If a text segment spans several lines, it will
+    # be stored in different text segments.
+    for segment in layout.text_anchor.text_segments:
+        start_index = int(segment.start_index)
+        end_index = int(segment.end_index)
+        response += text[start_index:end_index]
+    return response
+
+
+# [END documentai_process_form_document]
+# [END documentai_process_ocr_document]
