+++ documentai/snippets/list_processor_versions_sample.py
@@ -27,7 +27,7 @@ from google.cloud import documentai  # type: ignore
 def list_processor_versions_sample(
     project_id: str, location: str, processor_id: str
 ) -> None:
-    # You must set the api_endpoint if you use a location other than 'us'.
+    # You must set the `api_endpoint` if you use a location other than "us".
     opts = ClientOptions(api_endpoint=f"{location}-documentai.googleapis.com")
 
     client = documentai.DocumentProcessorServiceClient(client_options=opts)
