--- documentai/snippets/process_document_ocr_sample.py
@@ -1,181 +1,0 @@
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-# [START documentai_process_ocr_document]
-
-from __future__ import annotations
-
-from collections.abc import Sequence
-
-from google.api_core.client_options import ClientOptions
-from google.cloud import documentai  # type: ignore
-
-# TODO(developer): Uncomment these variables before running the sample.
-# project_id = 'YOUR_PROJECT_ID'
-# location = 'YOUR_PROCESSOR_LOCATION' # Format is 'us' or 'eu'
-# processor_id = 'YOUR_PROCESSOR_ID' # Create processor before running sample
-# processor_version = 'rc' # Refer to https://cloud.google.com/document-ai/docs/manage-processor-versions for more information
-# file_path = '/path/to/local/pdf'
-# mime_type = 'application/pdf' # Refer to https://cloud.google.com/document-ai/docs/file-types for supported file types
-
-
-def process_document_ocr_sample(
-    project_id: str,
-    location: str,
-    processor_id: str,
-    processor_version: str,
-    file_path: str,
-    mime_type: str,
-) -> None:
-    # Online processing request to Document AI
-    document = process_document(
-        project_id, location, processor_id, processor_version, file_path, mime_type
-    )
-
-    # For a full list of Document object attributes, please reference this page:
-    # https://cloud.google.com/python/docs/reference/documentai/latest/google.cloud.documentai_v1.types.Document
-
-    text = document.text
-    print(f"Full document text: {text}\n")
-    print(f"There are {len(document.pages)} page(s) in this document.\n")
-
-    for page in document.pages:
-        print(f"Page {page.page_number}:")
-        print_page_dimensions(page.dimension)
-        print_detected_langauges(page.detected_languages)
-        print_paragraphs(page.paragraphs, text)
-        print_blocks(page.blocks, text)
-        print_lines(page.lines, text)
-        print_tokens(page.tokens, text)
-
-        # Currently supported in version pretrained-ocr-v1.1-2022-09-12
-        if page.image_quality_scores:
-            print_image_quality_scores(page.image_quality_scores)
-
-
-def process_document(
-    project_id: str,
-    location: str,
-    processor_id: str,
-    processor_version: str,
-    file_path: str,
-    mime_type: str,
-) -> documentai.Document:
-    # You must set the api_endpoint if you use a location other than 'us'.
-    opts = ClientOptions(api_endpoint=f"{location}-documentai.googleapis.com")
-
-    client = documentai.DocumentProcessorServiceClient(client_options=opts)
-
-    # The full resource name of the processor version
-    # e.g. projects/{project_id}/locations/{location}/processors/{processor_id}/processorVersions/{processor_version_id}
-    # You must create processors before running sample code.
-    name = client.processor_version_path(
-        project_id, location, processor_id, processor_version
-    )
-
-    # Read the file into memory
-    with open(file_path, "rb") as image:
-        image_content = image.read()
-
-    # Load Binary Data into Document AI RawDocument Object
-    raw_document = documentai.RawDocument(content=image_content, mime_type=mime_type)
-
-    # Configure the process request
-    request = documentai.ProcessRequest(name=name, raw_document=raw_document)
-
-    result = client.process_document(request=request)
-
-    return result.document
-
-
-def print_page_dimensions(dimension: documentai.Document.Page.Dimension) -> None:
-    print(f"    Width: {str(dimension.width)}")
-    print(f"    Height: {str(dimension.height)}")
-
-
-def print_detected_langauges(
-    detected_languages: Sequence[documentai.Document.Page.DetectedLanguage],
-) -> None:
-    print("    Detected languages:")
-    for lang in detected_languages:
-        code = lang.language_code
-        print(f"        {code} ({lang.confidence:.1%} confidence)")
-
-
-def print_paragraphs(
-    paragraphs: Sequence[documentai.Document.Page.Paragraph], text: str
-) -> None:
-    print(f"    {len(paragraphs)} paragraphs detected:")
-    first_paragraph_text = layout_to_text(paragraphs[0].layout, text)
-    print(f"        First paragraph text: {repr(first_paragraph_text)}")
-    last_paragraph_text = layout_to_text(paragraphs[-1].layout, text)
-    print(f"        Last paragraph text: {repr(last_paragraph_text)}")
-
-
-def print_blocks(blocks: Sequence[documentai.Document.Page.Block], text: str) -> None:
-    print(f"    {len(blocks)} blocks detected:")
-    first_block_text = layout_to_text(blocks[0].layout, text)
-    print(f"        First text block: {repr(first_block_text)}")
-    last_block_text = layout_to_text(blocks[-1].layout, text)
-    print(f"        Last text block: {repr(last_block_text)}")
-
-
-def print_lines(lines: Sequence[documentai.Document.Page.Line], text: str) -> None:
-    print(f"    {len(lines)} lines detected:")
-    first_line_text = layout_to_text(lines[0].layout, text)
-    print(f"        First line text: {repr(first_line_text)}")
-    last_line_text = layout_to_text(lines[-1].layout, text)
-    print(f"        Last line text: {repr(last_line_text)}")
-
-
-def print_tokens(tokens: Sequence[documentai.Document.Page.Token], text: str) -> None:
-    print(f"    {len(tokens)} tokens detected:")
-    first_token_text = layout_to_text(tokens[0].layout, text)
-    first_token_break_type = tokens[0].detected_break.type_.name
-    print(f"        First token text: {repr(first_token_text)}")
-    print(f"        First token break type: {repr(first_token_break_type)}")
-    last_token_text = layout_to_text(tokens[-1].layout, text)
-    last_token_break_type = tokens[-1].detected_break.type_.name
-    print(f"        Last token text: {repr(last_token_text)}")
-    print(f"        Last token break type: {repr(last_token_break_type)}")
-
-
-def print_image_quality_scores(
-    image_quality_scores: documentai.Document.Page.ImageQualityScores,
-) -> None:
-    print(f"    Quality score: {image_quality_scores.quality_score:.1%}")
-    print("    Detected defects:")
-
-    for detected_defect in image_quality_scores.detected_defects:
-        print(f"        {detected_defect.type_}: {detected_defect.confidence:.1%}")
-
-
-def layout_to_text(layout: documentai.Document.Page.Layout, text: str) -> str:
-    """
-    Document AI identifies text in different parts of the document by their
-    offsets in the entirety of the document's text. This function converts
-    offsets to a string.
-    """
-    response = ""
-    # If a text segment spans several lines, it will
-    # be stored in different text segments.
-    for segment in layout.text_anchor.text_segments:
-        start_index = int(segment.start_index)
-        end_index = int(segment.end_index)
-        response += text[start_index:end_index]
-    return response
-
-
-# [END documentai_process_ocr_document]
