--- documentai/snippets/process_document_processor_version_sample.py
@@ -1,75 +1,0 @@
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-from typing import Optional
-
-# [START documentai_process_document_processor_version]
-
-from google.api_core.client_options import ClientOptions
-from google.cloud import documentai  # type: ignore
-
-# TODO(developer): Uncomment these variables before running the sample.
-# project_id = 'YOUR_PROJECT_ID'
-# location = 'YOUR_PROCESSOR_LOCATION' # Format is 'us' or 'eu'
-# processor_id = 'YOUR_PROCESSOR_ID' #  Create processor before running sample
-# processor_version_id = 'YOUR_PROCESSOR_VERSION_ID' # Processor version to use
-# file_path = '/path/to/local/pdf'
-# mime_type = 'application/pdf' # Refer to https://cloud.google.com/document-ai/docs/file-types for supported file types
-# field_mask = "text,entities,pages.pageNumber"  # Optional. The fields to return in the Document object.
-
-
-def process_document_processor_version_sample(
-    project_id: str,
-    location: str,
-    processor_id: str,
-    processor_version_id: str,
-    file_path: str,
-    mime_type: str,
-    field_mask: Optional[str] = None,
-) -> None:
-    # You must set the api_endpoint if you use a location other than 'us'.
-    opts = ClientOptions(api_endpoint=f"{location}-documentai.googleapis.com")
-
-    client = documentai.DocumentProcessorServiceClient(client_options=opts)
-
-    # The full resource name of the processor version
-    # e.g. projects/{project_id}/locations/{location}/processors/{processor_id}/processorVersions/{processor_version_id}
-    name = client.processor_version_path(
-        project_id, location, processor_id, processor_version_id
-    )
-
-    # Read the file into memory
-    with open(file_path, "rb") as image:
-        image_content = image.read()
-
-    # Load Binary Data into Document AI RawDocument Object
-    raw_document = documentai.RawDocument(content=image_content, mime_type=mime_type)
-
-    # Configure the process request
-    request = documentai.ProcessRequest(
-        name=name, raw_document=raw_document, field_mask=field_mask
-    )
-
-    result = client.process_document(request=request)
-
-    # For a full list of Document object attributes, please reference this page:
-    # https://cloud.google.com/python/docs/reference/documentai/latest/google.cloud.documentai_v1.types.Document
-    document = result.document
-
-    # Read the text recognition output from the processor
-    print("The document contains the following text:")
-    print(document.text)
-
-
-# [END documentai_process_document_processor_version]
