--- documentai/snippets/process_document_quality_sample.py
@@ -1,80 +1,0 @@
-# Copyright 2020 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-# [START documentai_process_quality_document]
-
-from google.api_core.client_options import ClientOptions
-from google.cloud import documentai  # type: ignore
-
-# TODO(developer): Uncomment these variables before running the sample.
-# project_id = 'YOUR_PROJECT_ID'
-# location = 'YOUR_PROCESSOR_LOCATION' # Format is 'us' or 'eu'
-# processor_id = 'YOUR_PROCESSOR_ID' # Create processor before running sample
-# file_path = '/path/to/local/pdf'
-# mime_type = 'application/pdf' # Refer to https://cloud.google.com/document-ai/docs/file-types for supported file types
-
-
-def process_document_quality_sample(
-    project_id: str, location: str, processor_id: str, file_path: str, mime_type: str
-) -> None:
-    # Online processing request to Document AI
-    document = process_document(
-        project_id, location, processor_id, file_path, mime_type
-    )
-
-    # Read the quality-specific information from the output from the
-    # Intelligent Document Quality Processor:
-    # https://cloud.google.com/document-ai/docs/processors-list#processor_doc-quality-processor
-    # OCR and other data is also present in the quality processor's response.
-    # Please see the OCR and other samples for how to parse other data in the
-    # response.
-    for entity in document.entities:
-        conf_percent = f"{entity.confidence:.1%}"
-        page_num = str(int(entity.page_anchor.page_refs[0].page) + 1)
-        print(f"\nPage {page_num} has a quality score of {conf_percent}")
-
-        for prop in entity.properties:
-            conf_percent = f"{prop.confidence:.1%}"
-            print(f"    * {prop.type_} score of {conf_percent}")
-
-
-def process_document(
-    project_id: str, location: str, processor_id: str, file_path: str, mime_type: str
-) -> documentai.Document:
-    # You must set the api_endpoint if you use a location other than 'us'.
-    opts = ClientOptions(api_endpoint=f"{location}-documentai.googleapis.com")
-
-    client = documentai.DocumentProcessorServiceClient(client_options=opts)
-
-    # The full resource name of the processor, e.g.:
-    # projects/project_id/locations/location/processor/processor_id
-    name = client.processor_path(project_id, location, processor_id)
-
-    # Read the file into memory
-    with open(file_path, "rb") as image:
-        image_content = image.read()
-
-    # Load Binary Data into Document AI RawDocument Object
-    raw_document = documentai.RawDocument(content=image_content, mime_type=mime_type)
-
-    # Configure the process request
-    request = documentai.ProcessRequest(name=name, raw_document=raw_document)
-
-    result = client.process_document(request=request)
-
-    return result.document
-
-
-# [END documentai_process_quality_document]
