+++ documentai/snippets/process_document_sample.py
@@ -12,20 +12,22 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from typing import Optional
 
 # [START documentai_process_document]
+# [START documentai_process_document_processor_version]
+from typing import Optional
 
 from google.api_core.client_options import ClientOptions
 from google.cloud import documentai  # type: ignore
 
 # TODO(developer): Uncomment these variables before running the sample.
-# project_id = 'YOUR_PROJECT_ID'
-# location = 'YOUR_PROCESSOR_LOCATION' # Format is 'us' or 'eu'
-# processor_id = 'YOUR_PROCESSOR_ID' # Create processor before running sample
-# file_path = '/path/to/local/pdf'
-# mime_type = 'application/pdf' # Refer to https://cloud.google.com/document-ai/docs/file-types for supported file types
+# project_id = "YOUR_PROJECT_ID"
+# location = "YOUR_PROCESSOR_LOCATION" # Format is "us" or "eu"
+# processor_id = "YOUR_PROCESSOR_ID" # Create processor before running sample
+# file_path = "/path/to/local/pdf"
+# mime_type = "application/pdf" # Refer to https://cloud.google.com/document-ai/docs/file-types for supported file types
 # field_mask = "text,entities,pages.pageNumber"  # Optional. The fields to return in the Document object.
+# processor_version_id = "YOUR_PROCESSOR_VERSION_ID" # Optional. Processor version to use
 
 
 def process_document_sample(
@@ -35,21 +37,29 @@ def process_document_sample(
     file_path: str,
     mime_type: str,
     field_mask: Optional[str] = None,
+    processor_version_id: Optional[str] = None,
 ) -> None:
-    # You must set the api_endpoint if you use a location other than 'us'.
+    # You must set the `api_endpoint` if you use a location other than "us".
     opts = ClientOptions(api_endpoint=f"{location}-documentai.googleapis.com")
 
     client = documentai.DocumentProcessorServiceClient(client_options=opts)
 
-    # The full resource name of the processor, e.g.:
-    # projects/{project_id}/locations/{location}/processors/{processor_id}
-    name = client.processor_path(project_id, location, processor_id)
+    if processor_version_id:
+        # The full resource name of the processor version, e.g.:
+        # `projects/{project_id}/locations/{location}/processors/{processor_id}/processorVersions/{processor_version_id}`
+        name = client.processor_version_path(
+            project_id, location, processor_id, processor_version_id
+        )
+    else:
+        # The full resource name of the processor, e.g.:
+        # `projects/{project_id}/locations/{location}/processors/{processor_id}`
+        name = client.processor_path(project_id, location, processor_id)
 
     # Read the file into memory
     with open(file_path, "rb") as image:
         image_content = image.read()
 
-    # Load Binary Data into Document AI RawDocument Object
+    # Load binary data
     raw_document = documentai.RawDocument(content=image_content, mime_type=mime_type)
 
     # Configure the process request
@@ -59,8 +69,8 @@ def process_document_sample(
 
     result = client.process_document(request=request)
 
-    # For a full list of Document object attributes, please reference this page:
-    # https://cloud.google.com/python/docs/reference/documentai/latest/google.cloud.documentai_v1.types.Document
+    # For a full list of `Document` object attributes, reference this page:
+    # https://cloud.google.com/document-ai/docs/reference/rest/v1/Document
     document = result.document
 
     # Read the text recognition output from the processor
@@ -68,4 +78,5 @@ def process_document_sample(
     print(document.text)
 
 
+# [END documentai_process_document_processor_version]
 # [END documentai_process_document]
