+++ documentai/snippets/quickstart_sample.py
@@ -13,52 +13,70 @@
 # limitations under the License.
 #
 
+# flake8: noqa
 
 # [START documentai_quickstart]
-
 
 from google.api_core.client_options import ClientOptions
 from google.cloud import documentai  # type: ignore
 
 # TODO(developer): Uncomment these variables before running the sample.
-# project_id = 'YOUR_PROJECT_ID'
-# location = 'YOUR_PROCESSOR_LOCATION' # Format is 'us' or 'eu'
-# processor_id = 'YOUR_PROCESSOR_ID' #  Create processor before running sample
-# file_path = '/path/to/local/pdf'
-# mime_type = 'application/pdf' # Refer to https://cloud.google.com/document-ai/docs/file-types for supported file types
+# project_id = "YOUR_PROJECT_ID"
+# location = "YOUR_PROCESSOR_LOCATION"  # Format is "us" or "eu"
+# processor_display_name = "YOUR_PROCESSOR_DISPLAY_NAME" # Must be unique per project, e.g.: "My Processor"
+# processor_type = "YOUR_PROCESSOR_TYPE" # Use `fetch_processor_types()` to get available processor types
+# file_path = "/path/to/local/pdf"
+# mime_type = "application/pdf"  # Refer to https://cloud.google.com/document-ai/docs/file-types for supported file types
 
 
 def quickstart(
-    project_id: str, location: str, processor_id: str, file_path: str, mime_type: str
-) -> None:
-    # You must set the api_endpoint if you use a location other than 'us'.
+    project_id: str,
+    location: str,
+    processor_display_name: str,
+    processor_type: str,
+    file_path: str,
+    mime_type: str,
+):
+    # You must set the `api_endpoint`if you use a location other than "us".
     opts = ClientOptions(api_endpoint=f"{location}-documentai.googleapis.com")
 
     client = documentai.DocumentProcessorServiceClient(client_options=opts)
 
-    # The full resource name of the processor, e.g.:
-    # projects/project_id/locations/location/processor/processor_id
-    name = client.processor_path(project_id, location, processor_id)
+    # The full resource name of the location, e.g.:
+    # `projects/{project_id}/locations/{location}`
+    parent = client.common_location_path(project_id, location)
+
+    # Create a Processor
+    processor = client.create_processor(
+        parent=parent,
+        processor=documentai.Processor(
+            display_name=processor_display_name, type_=processor_type
+        ),
+    )
+
+    # Print the processor information
+    print(f"Processor Name: {processor.name}")
 
     # Read the file into memory
     with open(file_path, "rb") as image:
         image_content = image.read()
 
-    # Load Binary Data into Document AI RawDocument Object
+    # Load binary data
     raw_document = documentai.RawDocument(content=image_content, mime_type=mime_type)
 
     # Configure the process request
-    request = documentai.ProcessRequest(name=name, raw_document=raw_document)
+    # `processor.name` is the full resource name of the processor, e.g.:
+    # `projects/{project_id}/locations/{location}/processors/{processor_id}`
+    request = documentai.ProcessRequest(name=processor.name, raw_document=raw_document)
 
     result = client.process_document(request=request)
 
-    # For a full list of Document object attributes, please reference this page:
-    # https://cloud.google.com/python/docs/reference/documentai/latest/google.cloud.documentai_v1.types.Document
+    # For a full list of `Document` object attributes, reference this page:
+    # https://cloud.google.com/document-ai/docs/reference/rest/v1/Document
     document = result.document
 
     # Read the text recognition output from the processor
     print("The document contains the following text:")
     print(document.text)
-
-
-# [END documentai_quickstart]
+    # [END documentai_quickstart]
+    return processor
