+++ documentai/snippets/quickstart_sample_test.py
@@ -16,25 +16,42 @@
 # flake8: noqa
 
 import os
+from uuid import uuid4
 
 from documentai.snippets import quickstart_sample
 
+from google.api_core.client_options import ClientOptions
+from google.cloud import documentai  # type: ignore
+
 location = "us"
 project_id = os.environ["GOOGLE_CLOUD_PROJECT"]
-processor_id = "90484cfdedb024f6"
+processor_display_name = f"test-processor-{uuid4()}"
+processor_type = "OCR_PROCESSOR"
 file_path = "resources/invoice.pdf"
 mime_type = "application/pdf"
 
 
 def test_quickstart(capsys):
-    quickstart_sample.quickstart(
+    processor = quickstart_sample.quickstart(
         project_id=project_id,
         location=location,
-        processor_id=processor_id,
+        processor_display_name=processor_display_name,
+        processor_type=processor_type,
         file_path=file_path,
         mime_type=mime_type,
     )
     out, _ = capsys.readouterr()
 
+    # Delete created processor
+    client = documentai.DocumentProcessorServiceClient(
+        client_options=ClientOptions(
+            api_endpoint=f"{location}-documentai.googleapis.com"
+        )
+    )
+    operation = client.delete_processor(name=processor.name)
+    # Wait for operation to complete
+    operation.result()
+
+    assert "Processor Name:" in out
     assert "text:" in out
     assert "Invoice" in out
